Cleanup `@internal` APIs from `presentation-components` barrel
